requirements.txt
streamlit
python-dotenv
langchain
langchain-community
docx2txt
PyPDF2
sentence-transformers
pinecone-client
google-generativeai